Support indexing from multiple data sources andd fix unit tests and queries

+++ omero_search_engine/cache_functions/elasticsearch/transform_data.py
@@ -70,6 +70,7 @@ def create_index(es_index, template):
                 es_index=es_index, error=str(ex)
             )
         )
+        raise ex
         return False
 
     return True
@@ -188,9 +189,10 @@ def delete_index(resource, es_index=None):
         return False
 
 
-def prepare_images_data(data, doc_type):
+def prepare_images_data(data, data_source, doc_type):
     data_record = [
         "id",
+        "data_source",
         "owner_id",
         "experiment",
         "group_id",
@@ -230,7 +232,10 @@ def prepare_images_data(data, doc_type):
             for rcd in data_record:
                 if rcd in ["mapvalue_name", "mapvalue_value"]:
                     continue
-                row_to_insert[rcd] = row[rcd]
+                elif rcd == "data_source":
+                    row_to_insert[rcd] = data_source
+                else:
+                    row_to_insert[rcd] = row[rcd]
 
             row_to_insert["key_values"] = []
             data_to_be_inserted[row["id"]] = row_to_insert
@@ -246,9 +251,10 @@ def prepare_images_data(data, doc_type):
     return data_to_be_inserted
 
 
-def prepare_data(data, doc_type):
+def prepare_data(data, data_source, doc_type):
     data_record = [
         "id",
+        "data_source",
         "owner_id",
         "group_id",
         "name",
@@ -277,7 +283,10 @@ def prepare_data(data, doc_type):
             for rcd in data_record:
                 if rcd in ["mapvalue_name", "mapvalue_value"]:
                     continue
-                row_to_insert[rcd] = row.get(rcd)
+                elif rcd == "data_source":
+                    row_to_insert[rcd] = data_source
+                else:
+                    row_to_insert[rcd] = row.get(rcd)
             row_to_insert["key_values"] = []
             data_to_be_inserted[row["id"]] = row_to_insert
         key_value = row_to_insert["key_values"]
@@ -292,7 +301,7 @@ def prepare_data(data, doc_type):
     return data_to_be_inserted
 
 
-def handle_file(file_name, es_index, cols, is_image, from_json):
+def handle_file(file_name, es_index, cols, is_image, data_source, from_json):
     co = 0
     search_omero_app.logger.info("Reading the csv file")
     if not from_json:
@@ -301,9 +310,9 @@ def handle_file(file_name, es_index, cols, is_image, from_json):
         df.columns = cols
         search_omero_app.logger.info("Prepare the data...")
         if not is_image:
-            data_to_be_inserted = prepare_data(df, es_index)
+            data_to_be_inserted = prepare_data(df, data_source, es_index)
         else:
-            data_to_be_inserted = prepare_images_data(df, es_index)
+            data_to_be_inserted = prepare_images_data(df, data_source, es_index)
         # print (data_to_be_inserted)
         search_omero_app.logger.info(len(data_to_be_inserted))
         with open(file_name + ".json", "w") as outfile:
@@ -344,7 +353,7 @@ def get_file_list(path_name):
     return f
 
 
-def insert_resource_data(folder, resource, from_json):
+def insert_resource_data(folder, resource, data_source, from_json):
     search_omero_app.logger.info(
         "Adding data to\
         {} using {}".format(
@@ -364,6 +373,7 @@ def insert_resource_data(folder, resource, from_json):
         is_image = True
         cols = [
             "id",
+            "data_source",
             "owner_id",
             "experiment",
             "group_id",
@@ -388,6 +398,7 @@ def insert_resource_data(folder, resource, from_json):
         if resource == "well":
             cols = [
                 "id",
+                "data_source",
                 "owner_id",
                 "group_id",
                 "mapvalue_name",
@@ -397,6 +408,7 @@ def insert_resource_data(folder, resource, from_json):
         else:
             cols = [
                 "id",
+                "data_source",
                 "owner_id",
                 "group_id",
                 "name",
@@ -422,20 +434,21 @@ def insert_resource_data(folder, resource, from_json):
         n = len(files_list)
         search_omero_app.logger.info("%s==%s == %s" % (f_con, fil, n))
         file_name = os.path.join(folder, fil)
-        handle_file(file_name, es_index, cols, is_image, from_json)
+        handle_file(file_name, es_index, cols, is_image, data_source, from_json)
         search_omero_app.logger.info("File: %s has been processed" % fil)
         try:
             with open(file_name + ".done", "w") as outfile:
                 json.dump(f_con, outfile)
-        except Exception:
+        except Exception as ex:
             search_omero_app.logger.info("Error .... writing Done file ...")
+            raise ex
         f_con += 1
 
 
 total_process = 0
 
 
-def get_insert_data_to_index(sql_st, resource, data_source):
+def get_insert_data_to_index(sql_st, resource, data_source, clean_index=True):
     """
     - Query the postgreSQL database server and get metadata (key-value pair)
     - Process the results data
@@ -446,11 +459,12 @@ def get_insert_data_to_index(sql_st, resource, data_source):
     """
     from datetime import datetime
 
-    delete_index(resource)
-    create_omero_indexes(resource)
+    if clean_index:
+        delete_index(resource)
+        create_omero_indexes(resource)
     sql_ = "select max (id) from %s" % resource
     res2 = search_omero_app.config.database_connectors[data_source].execute_query(sql_)
-    #res2 = search_omero_app.config["database_connector"].execute_query(sql_)
+    # res2 = search_omero_app.config["database_connector"].execute_query(sql_)
     max_id = res2[0]["max"]
     if not max_id:
         return
@@ -467,7 +481,7 @@ def get_insert_data_to_index(sql_st, resource, data_source):
     global total_process
     total_process = len(vals)
     # Determine the number of processes inside the multiprocessing pool,
-    # i.e the number of allowed processors to run at the same time
+    # i.e. the number of allowed processors to run at the same time
     # It depends on the number of the processors that the hosting machine has
     no_processors = search_omero_app.config.get("NO_PROCESSES")
     if not no_processors:
@@ -487,11 +501,14 @@ def get_insert_data_to_index(sql_st, resource, data_source):
         counter_val = manager.Value("i", 0)
         func = partial(processor_work, lock, counter_val)
         # map the data which will be consumed by the processes inside the pool
-        res = pool.map(func, vals)
+        res = pool.map(func, vals)  # noqa
         search_omero_app.logger.info(cur_max_id)
         delta = str(datetime.now() - start_time)
         search_omero_app.logger.info("Total time=%s" % delta)
-        #print(res)
+        # print(res)
+    except Exception as ex:
+        print("Error is : %s" % ex)
+        raise ex
     finally:
         pool.close()
 
@@ -513,6 +530,9 @@ def processor_work(lock, global_counter, val):
     try:
         lock.acquire()
         global_counter.value += 1
+    except Exception as ex:
+        print("Error is %s" % ex)
+        raise ex
     finally:
         lock.release()
     whereclause = " where %s.id < %s and %s.id >= %s" % (
@@ -529,18 +549,18 @@ def processor_work(lock, global_counter, val):
     conn = search_omero_app.config.database_connectors[data_source]
     results = conn.execute_query(mod_sql)
     search_omero_app.logger.info("Processing the results...")
-    process_results(results, resource, lock)
+    process_results(results, resource, data_source, lock)
     average_time = (datetime.now() - st) / 2
     search_omero_app.logger.info("Done")
     search_omero_app.logger.info("elpased time:%s" % average_time)
 
 
-def process_results(results, resource, lock=None):
+def process_results(results, resource, data_source, lock=None):
     df = pd.DataFrame(results).replace({np.nan: None})
-    insert_resource_data_from_df(df, resource, lock)
+    insert_resource_data_from_df(df, resource, data_source, lock)
 
 
-def insert_resource_data_from_df(df, resource, lock=None):
+def insert_resource_data_from_df(df, resource, data_source, lock=None):
     if resource == "image":
         is_image = True
     else:
@@ -548,9 +568,9 @@ def insert_resource_data_from_df(df, resource, lock=None):
     es_index = resource_elasticsearchindex.get(resource)
     search_omero_app.logger.info("Prepare the data...")
     if not is_image:
-        data_to_be_inserted = prepare_data(df, es_index)
+        data_to_be_inserted = prepare_data(df, data_source, es_index)
     else:
-        data_to_be_inserted = prepare_images_data(df, es_index)
+        data_to_be_inserted = prepare_images_data(df, data_source, es_index)
     # print (data_to_be_inserted)
     search_omero_app.logger.info(len(data_to_be_inserted))
 
@@ -636,7 +656,7 @@ def insert_plate_data(folder, plate_file):
 
 
 def save_key_value_buckets(
-    resource_table_=None, data_source=None, re_create_index=False, only_values=False
+    resource_table_=None, data_source=None, clean_index=False, only_values=False
 ):
     """
     Query the database and get all available keys and values for
@@ -645,12 +665,12 @@ def save_key_value_buckets(
     It will use multiprocessing pool to use parallel processing
 
     """
-    if data_source==None:
-        return
+    if data_source is None:
+        return "No data source is provided"
     es_index = "key_value_buckets_information"
     es_index_2 = "key_values_resource_cach"
 
-    if re_create_index:
+    if clean_index:
         if not only_values:
             search_omero_app.logger.info(
                 "Try to delete if exist:  %s " % delete_es_index(es_index)
@@ -699,6 +719,7 @@ def save_key_value_buckets(
             name_results = [
                 {
                     "id": res["id"],
+                    # "data_source": data_source,
                     "description": res["description"],
                     "name": res["name"],
                     "no_images": res["no_images"],
@@ -706,7 +727,10 @@ def save_key_value_buckets(
                 for res in name_result
             ]
 
-        push_keys_cache_index(resource_keys, resource_table, es_index_2, name_results)
+        push_keys_cache_index(
+            resource_keys, resource_table, data_source, es_index_2, name_results
+        )
+        print(type(resource_keys), type(resource_table), es_index_2, type(name_results))
         if only_values:
             continue
         search_omero_app.logger.info(
@@ -718,7 +742,9 @@ def save_key_value_buckets(
         # prepare the data which will be consumed by the processes
         # inside the multiprocessing Pool
         for key in resource_keys:
-            vals.append((key, resource_table, es_index, len(resource_keys)))
+            vals.append(
+                (key, resource_table, es_index, len(resource_keys), data_source)
+            )
         # determine the number of processes inside the process pool
         no_processors = search_omero_app.config.get("NO_PROCESSES")
         if not no_processors:
@@ -747,6 +773,7 @@ def save_key_value_buckets_process(lock, global_counter, vals):
     resource_table = vals[1]
     es_index = vals[2]
     total = vals[3]
+    data_source = vals[4]
     wrong_keys = {}
     try:
         lock.acquire()
@@ -760,7 +787,10 @@ def save_key_value_buckets_process(lock, global_counter, vals):
             % (global_counter.value, total)
         )
         search_omero_app.logger.info("Checking {key}".format(key=key))
-        data_to_be_pushed = get_buckets(key, resource_table, es_index, lock)
+        data_to_be_pushed = get_buckets(
+            key, data_source, resource_table, es_index, lock
+        )
+
         actions = []
         search_omero_app.logger.info(
             "data_to_be_pushed:\
@@ -776,6 +806,7 @@ def save_key_value_buckets_process(lock, global_counter, vals):
             search_omero_app.logger.info(helpers.bulk(es, actions))
         except Exception as e:
             search_omero_app.logger.info("Error:  %s" % str(e))
+            raise e
             # raise e
         finally:
             lock.release()
@@ -785,6 +816,7 @@ def save_key_value_buckets_process(lock, global_counter, vals):
             Error:%s "
             % (global_counter.value, str(e))
         )
+        raise e
         if wrong_keys.get(resource_table):
             wrong_keys[resource_table] = wrong_keys[resource_table].append(key)
         else:
@@ -798,17 +830,22 @@ def get_keys(res_table, data_source):
            annotation_mapvalue.annotation_id".format(
         res_table=res_table
     )
-    results = search_omero_app.config.database_connectors[data_source].execute_query(sql)
-    #results = search_omero_app.config["database_connector"].execute_query(sql)
+    results = search_omero_app.config.database_connectors[data_source].execute_query(
+        sql
+    )
+    # results = search_omero_app.config["database_connector"].execute_query(sql)
     results = [res["name"] for res in results]
     return results
 
 
-def push_keys_cache_index(results, resource, es_index, resourcename=None):
+def push_keys_cache_index(
+    results, resource, data_resource, es_index, resourcename=None
+):
     row = {}
     row["name"] = results
     row["doc_type"] = es_index
     row["resource"] = resource
+    row["data_resource"] = data_resource
     if resourcename:
         row["resourcename"] = resourcename
 
@@ -819,20 +856,23 @@ def push_keys_cache_index(results, resource, es_index, resourcename=None):
     search_omero_app.logger.info(helpers.bulk(es, actions))
 
 
-def get_buckets(key, resourcse, es_index, lock=None):
+def get_buckets(key, data_source, resourcse, es_index, lock=None):
     try:
         lock.acquire()
-        res = get_all_values_for_a_key(resourcse, key)
+        res = get_all_values_for_a_key(resourcse, data_source, key)
+    except Exception as ex:
+        print("Error %s" % ex)
+        raise ex
     finally:
         lock.release()
     search_omero_app.logger.info(
         "number of bucket: %s" % res.get("total_number_of_buckets")
     )
-    data_to_be_pushed = prepare_bucket_index_data(res, resourcse, es_index)
+    data_to_be_pushed = prepare_bucket_index_data(res, resourcse, data_source, es_index)
     return data_to_be_pushed
 
 
-def prepare_bucket_index_data(results, res_table, es_index):
+def prepare_bucket_index_data(results, res_table, data_source, es_index):
     data_to_be_inserted = []
     for result in results.get("data"):
         row = {}
@@ -844,6 +884,7 @@ def prepare_bucket_index_data(results, res_table, es_index):
         row["Value"] = result["Value"]
         row["items_in_the_bucket"] = result["Number of %ss" % res_table]
         row["total_buckets"] = results["total_number_of_buckets"]
+        row["data_source"] = data_source
         row["total_items_in_saved_buckets"] = results["total_number"]
         row["total_items"] = results["total_number_of_%s" % res_table]
     return data_to_be_inserted
